add shared ui
